feat(develop): 添加幂等注解，让接口具有幂等性

+++ personal-blog-common/src/main/java/com/gewuyou/blog/common/vo/ConditionVO.java
@@ -29,23 +29,23 @@ public class ConditionVO {
     @Schema(name = "keywords", description = "搜索内容", type = "String")
     private String keywords;
 
-    @Schema(name = "id", description = "分类id", type = "Integer")
-    private Integer categoryId;
+    @Schema(name = "id", description = "分类id", type = "Long")
+    private Long categoryId;
 
     @Schema(name = "tagId", description = "标签id", type = "Integer")
-    private Integer tagId;
+    private Long tagId;
 
     @Schema(name = "albumId", description = "相册id", type = "Integer")
     private Integer albumId;
 
-    @Schema(name = "loginType", description = "登录类型", type = "Integer")
-    private Integer loginType;
+    @Schema(name = "loginType", description = "登录类型", type = "Byte")
+    private Byte loginType;
 
     @Schema(name = "type", description = "类型", type = "Integer")
     private Integer type;
 
-    @Schema(name = "status", description = "状态", type = "Integer")
-    private Integer status;
+    @Schema(name = "status", description = "状态", type = "Byte")
+    private Byte status;
 
     @Schema(name = "startTime", description = "开始时间", type = "LocalDateTime")
     private LocalDateTime startTime;
@@ -54,14 +54,14 @@ public class ConditionVO {
     private LocalDateTime endTime;
 
     @Schema(name = "isDelete", description = "是否删除", type = "Integer")
-    private Integer isDelete;
+    private Byte isDelete;
 
     @Schema(name = "isReview", description = "是否审核", type = "Integer")
-    private Integer isReview;
+    private Byte isReview;
 
     @Schema(name = "isTop", description = "是否置顶", type = "Integer")
-    private Integer isTop;
+    private Byte isTop;
 
     @Schema(name = "isFeatured", description = "是否推荐", type = "Integer")
-    private Integer isFeatured;
+    private Byte isFeatured;
 }
